ai_service/backend/requirements.txt
fastapi==0.115.13
sqlmodel==0.0.24
# gunicorn==23.0.0
redis==6.2.0
# For running Postgres Database
psycopg2-binary==2.9.10 

python-dotenv==1.0.1

pytest==8.3.5

# Langchain
langchain==0.3.25
# langchain-ollama==0.3.3
langchain-community==0.3.24
langchain-milvus==0.1.10
langchain-google-genai==2.1.5

langgraph==0.4.8
langgraph-cli[inmem]==0.2.12

unstructured==0.16.21
networkx==3.4.2
openpyxl==3.1.5

beautifulsoup4==4.13.3